chore(payments-server): move AppLocalizationProvider into own file
and replace proptypes with typescript types
- fixes #3762
